+++ src/models/decoder_only_pipeline.cpp
@@ -243,7 +243,7 @@ void DecoderOnlyPipelineState::UpdateInputsOutputs(const RoamingArray<int32_t>& 
                                                    RoamingArray<int32_t> beam_indices, int total_length) {
   input_ids_.Update(next_tokens);
   size_t new_length = input_ids_.GetShape()[1];
-  position_inputs_.Update(next_tokens, total_length, new_length);
+  position_inputs_.Update(next_tokens, total_length, static_cast<int>(new_length));
   if (kv_cache_) kv_cache_->Update(beam_indices.GetCPU(), total_length);
   logits_.Update(next_tokens, new_length);
 }
