+++ src/models/logits.cpp
@@ -191,8 +191,8 @@ RoamingArray<float> Logits::Get() {
 
 #pragma warning(pop)
 
-void Logits::Update(const RoamingArray<int32_t>& next_tokens, int new_kv_length) {
-  if (output_raw_.get()->GetTensorTypeAndShapeInfo()->GetShape()[1] == new_kv_length && new_kv_length == 1) {
+void Logits::Update(const RoamingArray<int32_t>& next_tokens, size_t new_kv_length) {
+  if (static_cast<size_t>(output_raw_.get()->GetTensorTypeAndShapeInfo()->GetShape()[1]) == new_kv_length && new_kv_length == 1) {
     return;
   }
 
@@ -205,10 +205,10 @@ void Logits::Update(const RoamingArray<int32_t>& next_tokens, int new_kv_length)
       if (next_token != model_.config_->model.pad_token_id)
         break;
     }
-    input_sequence_lengths[b] = token_index + 1;
+    input_sequence_lengths[b] = static_cast<int>(token_index + 1);
   }
 
-  if (output_raw_.get()->GetTensorTypeAndShapeInfo()->GetShape()[1] == new_kv_length) {
+  if (static_cast<size_t>(output_raw_.get()->GetTensorTypeAndShapeInfo()->GetShape()[1]) == new_kv_length) {
     return;
   }
